Organized components in project folder to 3 seperate files
